Fix crucial download link pocketbase
At university some Fireship io fans and I (colleagues) tried to download pocketbase, but without success.

We tried to follow the guide, but we encounter an error on link: pocketbase.io, so with this PR we want to encourage new devs to implement with one click access "pocketbase" site to download, and allow other devs to implement as fast as they can the tool, We think this is an amazing and beautiful tool, and if a team of 6 devs are frustrated to did accomplish using it! We want to add this way, to easy, fast and secure implement 1 command go to site to download pocketbase on their machines! Thanks to all team to make projects like this!